ArtShah/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
automium==0.2.6
automium-web==0.1.1
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.1
cryptography==41.0.1
cycler==0.12.1
distlib==0.3.6
Django==3.2.13
django-auth-adfs==1.11.6
django-email-users==1.1.0
django-filter==22.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
filelock==3.8.0
fonttools==4.43.1
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
legacy==0.1.6
login==0.0.6
magic-filter==1.0.12
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.8.0
multidict==6.0.4
numpy==1.26.1
packaging==23.2
phonenumbers==8.13.10
Pillow==9.4.0
platformdirs==2.5.2
psycopg==3.1.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2022.7.1
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.16.6
Werkzeug==2.2.3
yarl==1.9.2
psycopg2==2.9.9